+++ app/filters/agilibox/small_data/filter_strategy_by_key_value.rb
@@ -2,6 +2,7 @@ class Agilibox::SmallData::FilterStrategyByKeyValue < ::Agilibox::SmallData::Fil
   attr_reader :key
 
   def initialize(key = nil)
+    super()
     @key = key
   end