+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Next version
+
+## v1.9.6
 - Fix info helper with false
 - Update Rubocop + fix offences